requirements.txt
pip
praw
discord.py